Remove fzf-related settings [b]

+++ setup.sh
@@ -10,8 +10,10 @@ fi
 
 dir="zsh/$1"
 
-rm ~/.fzf.zsh
-ln -s $PWD/$dir/.fzf.zsh ~/.fzf.zsh
+if [ "$1" = "a" ]; then
+  rm ~/.fzf.zsh
+  ln -s "$PWD/$dir/.fzf.zsh" ~/.fzf.zsh
+fi
 
 rm ~/.p10k.zsh
 ln -s $PWD/$dir/.p10k.zsh ~/.p10k.zsh
